backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0

# Scheduler
apscheduler==3.10.4

# Plaid SDK
plaid-python==29.0.0

# Utilities
pydantic==2.10.0
pydantic-settings==2.6.0
cryptography==44.0.0
httpx==0.28.0

# ID generation
cuid2==2.0.1
